backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pocketflow>=0.0.3
python-dotenv>=1.0.1
requests>=2.32.3
openai>=1.46.0
faiss-cpu>=1.8.0
numpy>=1.24.0
duckduckgo-search>=8.0.0
pydantic>=2.8.2
